requirements.txt
python-telegram-bot
pythonping